Fix voor WAD_Interface issue #40: - bij uploaden van een nieuwe config wordt gecheckt of de file al bestaat. - bij updaten van een config wordt de oude config verwijderd alvorens   de nieuwe config wordt gekopieerd (hoeft niet perse dezelfde filename   te hebben). - bij deleten van een config wordt nu ook de config verwijderd.

+++ website/WAD-IQC/database/iqc/new_analysemodule_cfg.php
@@ -4,6 +4,16 @@ require("../globals.php") ;
 require("./common.php") ;
 require("./php/includes/setup.php");
 
+// add new cfg : GET pk = 0
+// na toevoegen: GET pk = 0
+//               $_POST['action'] = "Add"
+// update cfg  : GET pk = pk van cfg in analysemodule_cfg tabel
+// na updaten  : GET pk = pk van cfg in analysemodule_cfg tabel
+//               $_POST['action'] = "Update"
+// delete cfg  : GET pk = -1 (via FORM action in POST)
+//               $_POST['analysemodule_cfg[pk]'] = "on"  (pk van te deleten cfg)
+//               $_POST['action'] = ""
+
 $pk=$_GET['pk'];
 
 
@@ -38,16 +48,16 @@ if (mysqli_connect_errno()) {
 
 
 
-
+// Aanroep na toevoegen (pk=0) of na updaten (pk>0)van config
 if(!empty($_POST['action']))
 {
  
   
-  $error    = $_FILES['uploadedfile']['error'];
+  $error    = $_FILES['uploadedfile']['error']; // 0 = success, 4 = no file uploaded
   $description=$_POST['description'];
   
   
-
+  // add new analysemodule cfg
   if ($pk==0)
   { 
     $filename=basename( $_FILES['uploadedfile']['name']);
@@ -55,6 +65,11 @@ if(!empty($_POST['action']))
     $filepath = $filepath_root.basename( $_FILES['uploadedfile']['name']); 
 
     $target_path=__DIR__ . '/../../../' . $filepath;
+
+    if (file_exists($target_path)) {
+       echo "Config \"$filename\" already exists! Please rename config and try to upload again, or click on config filename to update the module.";
+       exit();
+    }
 
     if ( move_uploaded_file($_FILES['uploadedfile']['tmp_name'], $target_path) )
     {
@@ -74,9 +89,10 @@ if(!empty($_POST['action']))
       exit() ;
     }
   }
+  // update analysemodule cfg
   if ($pk>0)
   { 
-    if ($error==4)    
+    if ($error==4)    // "no file uploaded" -> alleen omschrijving aanpassen
     {
       if (!($link->query(sprintf($update_Stmt1,$description,$pk))))  
       {
@@ -87,6 +103,24 @@ if(!empty($_POST['action']))
     }
     if ($error==0)    
     {
+       // delete eerst de oude config
+       if (!($result_analysemodule_cfg=$link->query(sprintf($select_Stmt,$pk)))) {
+         DisplayErrMsg(sprintf("Error in executing %s stmt", $subject_Stmt)) ;
+         DisplayErrMsg(sprintf("error: %s", $link->error)) ;
+         exit() ;
+       }
+       $field_analysemodule_cfg = $result_analysemodule_cfg->fetch_object();
+       $filename_db=$field_analysemodule_cfg->filename;
+       $filepath_db=$field_analysemodule_cfg->filepath;
+       $result_analysemodule_cfg->close();
+       $target_folder_db =__DIR__ . '/../../../' . $filepath_db;
+       $target_path_db = $target_folder_db . $filename_db;
+       if (!unlink($target_path_db))
+       {
+          echo ("Error deleting old config \"$filename_db\"");
+          exit();
+       }
+       // kopieer de nieuwe upload naar de analysemodule_cfg folder
        $filename=basename( $_FILES['uploadedfile']['name']);
        $filepath_root="WAD-IQC/uploads/analysemodule_cfg/";
        $filepath = $filepath_root.basename( $_FILES['uploadedfile']['name']); 
@@ -128,7 +162,7 @@ if (!empty($_POST['action']))
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
-// if it will get to here it is either the first time or it returned from add/modify picture
+// if it will get to here it is either the first time or it returned from add/modify config
 ///////////////////////////////////////////////////////////////////////////////////////////
 
 
@@ -141,9 +175,9 @@ if ($pk==-1)         //delete
   $limit=0;
   if (!empty($_POST['analysemodule_cfg']))
   {
-    $analysemodule_cfg=$_POST['analysemodule_cfg'];
-    $analysemodule_cfg_ref_key=array_keys($analysemodule_cfg);
-    $limit=sizeof($analysemodule_cfg_ref_key);
+    $analysemodule_cfg=$_POST['analysemodule_cfg'];            // array vd vorm { [pk1] = "on", [pk2] = "on", .. }
+    $analysemodule_cfg_ref_key=array_keys($analysemodule_cfg); // array met aangevinkte pk's
+    $limit=sizeof($analysemodule_cfg_ref_key);                 // aantal aangevinkte configs
   } 
   $i=0;
 
@@ -151,34 +185,25 @@ if ($pk==-1)         //delete
   {
     if ($analysemodule_cfg[$analysemodule_cfg_ref_key[$i]]=='on')
     {
-    
+      // haal config filename op uit DB op basis van pk
       if (!($result_analysemodule_cfg= $link->query(sprintf($select_Stmt,$analysemodule_cfg_ref_key[$i])))) {
       DisplayErrMsg(sprintf("Error in executing %s stmt", $subject_Stmt)) ;
       DisplayErrMsg(sprintf("error: %s", $link->error)) ;
       exit() ;
       }
       $field_analysemodule_cfg = $result_analysemodule_cfg->fetch_object();
+      $filename=$field_analysemodule_cfg->filename;
       $filepath=$field_analysemodule_cfg->filepath;
       $result_analysemodule_cfg->close();
 
-      if (!($result_analysemodule_cfg= $link->query(sprintf($select_Stmt1,$filepath)))) {
-      DisplayErrMsg(sprintf("Error in executing %s stmt", $subject_Stmt)) ;
-      DisplayErrMsg(sprintf("error: %s", $link->error)) ;
-      exit() ;
-      }
-      $counter=0;
-      while ($field_analysemodule_cfg = $result_analysemodule_cfg->fetch_object() )
+      $target_folder =__DIR__ . '/../../../' . $filepath;
+      $target_path = $target_folder . $filename;
+      if (!unlink($target_path))
       {
-        $counter++;
+         echo ("Error deleting \"$filename\"");
+         exit();
       }
-      $result_analysemodule_cfg->close();
-      if ($counter==1) //only 1 row that contains filepath
-      {
-        $target_path=$target_path.$filepath;
-        //printf("target=%s",$target_path);
-        //exit();
-        unlink($target_path);
-      } 
+
       if (!($result_analysemodule_cfg= $link->query(sprintf($del_analysemodule_cfg_Stmt,$analysemodule_cfg_ref_key[$i])))) {
       DisplayErrMsg(sprintf("Error in executing %s stmt", sprintf($del_analysemodule_cfg_Stmt,$analysemodule_cfg_ref_key[$i]) ) ) ;
       DisplayErrMsg(sprintf("error: %s", $link->error)) ;
@@ -248,15 +273,3 @@ if (mysqli_connect_errno()) {
 
    
 ?>
-
-
-
-
-
-
-
-
-
-
-
-
